// Code/src/app/shared/model/RegularClientModel.ts
import {Contact_info} from '../../shared/model/Contact_info';
export class RegularClientModel {
    id: number;
    branchId: number;
    clientCode: string;
    clientName: string;
    clientGroupId: number;
    pan: string;
    gstIN: string;
    address: string;
    cityId: number;
    pinCode: number;
    StateId: number;
    phoneNo: number;
    mobileNo: number;
    emailAlert: boolean;
    smsAlert: boolean;
    RegularClient_Contact_info:Contact_info[];
    deliveryAgainstAsCnr: string;
    deliveryAgainstAsCne: string;
    ledgerName: string;
    days: string;
    amount: string;
    interest: string;
    creditGraceDays : number;
    bankId: number;
    ifsc:number;
    accountNo: string;
    chequeInTheNameOf: string;
    companyNo: number;
    policyNo: number;
    validFromDate: Date;
    validToDate: Date;
    insuranceAmount: string;
    isMarineInsured: boolean;
    isGodownInsured: boolean;
    paymentMode: string;
    bookingType: string;
    deliveryType: string;
    multipleBillingBranches: string;
    branchCode:string;
    remark: string;
}